Realized that testing most modules without the physical jetson is impossible
Because the missing peripherals will always throw error's it's impossible to test modules like the pmc or apriltags on a docker container at this time.

The only container that shows promise in testing is sandbox and maybe simulator.

I need to do more testing to really figure out how docker works.