Update AVR software documentation
I'm trying my best here, but everything is up in the air (haha). There's no way to write a definitive guide that will work for every future team. The best I can do is write documentation for how things are right now, and advise the reader to learn how these things work so that they can be self-reliant for instructions.