Add network connection box to GUI
Created a new groupbox on the connections tab of the GUI. This groupbox is dedicated to simplifying the process of connecting to the AVR's wifi. It has a config option and displays the result of the connection attempt.

I have noticed that the displayed values don't always line up with the actual outcome.