Refactor manual control inputs to avoid reliance on Sandbox test commands
Refactor the manual control section of the GUI to operate on it's own movement commands, instead of relying on commands that are translated via the sandbox. This should slightly improve response time and avoid a major breaking point for manual control.
Thus, I removed the applicable testing commands from the sandbox.
I also made the armed status display on change to "armed" instead of "armed & dangerous", since that took up too much space.